+++ NEWS.md
@@ -9,6 +9,7 @@
     - `as_ggplot()`: Transform the output of `gridExtra::arrangeGrob()` and `gridExtra::grid.arrange()` to a an object of class ggplot.
     - `ggtexttable()`: to draw a textual table.
     - `ggparagraph()`: to draw a paragraph of text.
+    - fill_palette() and color_palette() to change the fill and color palette, respectively.
     
 - New theme function `theme_transparent()` to create a ggplot with transparent background.